Packaged go-service-example, which is an example HTTP go service for file storage.
Created a nix package for the go binary that allows file uploads and downloads.
The nix package is stored at pkgs/custom/go-service-example. Used custom folder
because it is not intended to be ever merged in the mainline nixpkgs.

The package build definition is using fetchFromGitHub to get the code using a
specific revition, because tags are not yet available for development version.

The service defintion is located at ./services/custom/go-service-examples. It
provides 3 configuration options, enable, dataDirectory and host. The service
is using systemd for starting and managing the binary.